// apps/web/src/api/generated/schemas/posMovementItem.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { PosMovementItemPosProvider } from './posMovementItemPosProvider';
import type { PosMovementItemPosTransactionId } from './posMovementItemPosTransactionId';
import type { PosMovementItemInstallment } from './posMovementItemInstallment';
import type { PosMovementItemErrorMessage } from './posMovementItemErrorMessage';
import type { PosMovementItemSaleId } from './posMovementItemSaleId';
import type { PosMovementItemPatientName } from './posMovementItemPatientName';

export interface PosMovementItem {
  id: string;
  date: string;
  amount: number;
  status: string;
  posProvider?: PosMovementItemPosProvider;
  posTransactionId?: PosMovementItemPosTransactionId;
  installment?: PosMovementItemInstallment;
  errorMessage?: PosMovementItemErrorMessage;
  saleId?: PosMovementItemSaleId;
  patientName?: PosMovementItemPatientName;
}